Implement new option classes
Finish implementing the new option classes, and finish refactoring all
of the code in option.cpp to use the new classes, as well as to use more
C++ (particularly std::string and std::ifstream) where it makes sense to
do so. This gets us to the point that option.cpp compiles, though other
sources still need some tweaking.