Start implementing new front-end tests
Implement new test runner, using a new FrontEndTest type added to the
framework. This consumes YAML as the test specifications and is focused
on validating the output (whether stdout, stderr, or a file written to
disk) of uncrustify. In particular, the goal is to eventually replace
test_cli_options.py with something more in line with the formatting
tests.

Also, start porting some of the existing test_cli_options.py tests.

The plan is to have yet another test type specifically for parsing
tests (with the intent that we will start adding many of these). Some
functions and types have been introduced toward this end, but are not
yet fully implemented.